fix a bug

+++ trainers/trainer.py
@@ -506,8 +506,8 @@ class _Trainer:
 
         score_mat = self.evaluator.compare_features_symmetry_y(features)
         N = score_mat.size(0)
-        indices_i = torch.Tensor([i for i in range(N)]).expand(N, N).contiguous().view(-1)
-        indices_j = torch.Tensor([i for i in range(N)]).expand(N, N).t().contiguous().view(-1)
+        indices_i = torch.Tensor([i for i in range(N)]).to(dtype=torch.long).expand(N, N).contiguous().view(-1)
+        indices_j = torch.Tensor([i for i in range(N)]).to(dtype=torch.long).expand(N, N).t().contiguous().view(-1)
 
         width = 32
         height = 64
